feat: verificacao da foto ASCII, exibicao da foto ASCII do candidato quando vota e pequeno ajuste no banco

+++ sistema/candidato.py
@@ -31,25 +31,85 @@ def gestao_candidatos(conn):
                             fotoASCII = str(input('Deseja adicionar foto ASCII para o candidato (s/n): ')).lower()
                             while fotoASCII != 's' and fotoASCII != 'n':
                                 print('Digite somente "s" ou "n"')
-                                fotoASCII = input('Deseja adicionar foto ASCII (s/n): ').lower()
+                                fotoASCII = str(input('Deseja adicionar foto ASCII para o candidato (s/n): ')).lower()
                                 
                             asciiFOTO = None
+                            pastaASCII = None
+                            
                             if fotoASCII == 's':
-                                os.system('cls')
-                                print("\nCole a arte ASCII (digite FIM para terminar):\n")
                                 
-                                linhas = []
-                                while True:
-                                    linha = input()
-                                    if linha == 'FIM':
-                                        break
-                                    linhas.append(linha)
-                                
-                                asciiFOTO = '\n'.join(linhas)
-                                pastaASCII = f'ascii/ascii_candidato{number}.txt'
-                                
-                                with open(pastaASCII, 'w', encoding='utf-8') as f:
-                                    f.write(asciiFOTO)
+                                repetir = 's'
+                                while repetir == 's':   
+                                    os.system('cls')
+                                    print("\nCole a arte ASCII (digite FIM para terminar):\n")
+                                    
+                                    linhas = []
+                                    linha = ''
+                                    while linha != 'FIM' :
+                                        linha = input()
+                                        if linha != 'FIM':
+                                            linhas.append(linha)
+                                    
+                                    '''
+                                        nesses if sao validacoes simples para ver se o cara nao ta adicionando qualquer coisa
+                                        1. verifico o numero de linhas se for muito pequeno eu ja invalido
+                                        2. verifico o numero de colunas para ver se o desenho nao vai ficar muito estreito, se for
+                                        muito pequeno ja invalido tbm
+                                        3. depois eu verifico se o cara ta digitando pelo menos 100 caracteres
+                                        4. peço pra confirmar e salvo na pasta ascii com o numero dele, ai no banco, eu salvo o
+                                        caminho da pasta, para ficar associado a foto com o candidato
+                                    '''
+                                    
+                                    if len(linhas) < 5:  
+                                        print('\nASCII muito pequeno!')
+                                        input('Pressione ENTER para tentar novamente')
+                                        continue
+                                    
+                                    maior_linha = 0
+                                    for linha in linhas:
+                                        tamanho = len(linha)
+                                        
+                                        if tamanho > maior_linha:
+                                            maior_linha = tamanho
+                                    
+                                    if maior_linha < 20:
+                                        print('\nASCII muito pequeno!')
+                                        input('Pressione ENTER para tentar novamente')
+                                        continue
+                                    
+                                    
+                                    contador = 0
+                                    for linha in linhas:
+                                        for caracter in linha:
+                                            if caracter != ' ':
+                                                contador +=1 
+                                    
+                                    if contador < 100:
+                                        print('\nASCII inválido')
+                                        input('Pressione ENTER para tentar novamente')
+                                        continue
+                                    
+                                    os.system('cls')
+                                    print('\nPré Vizualizacao')
+                                    print('\n'.join(linhas))
+                                    
+                                    confirmar = str(input('\nConfirmar ASCII (s/n): ')).lower()
+                                    while confirmar != 's' and confirmar != 'n':
+                                        print('Digite somente "s" ou "n"')
+                                        confirmar = input('Deseja adicionar foto ASCII (s/n): ').lower()
+
+                                    if confirmar != 's':
+                                        print('\nRefaça o ASCII')
+                                        input('Pressione ENTER')
+                                        continue
+                                        
+                                    asciiFOTO = '\n'.join(linhas)
+                                    pastaASCII = f'ascii/ascii_candidato{number}.txt'
+
+                                    with open(pastaASCII, 'w', encoding='utf-8') as f:
+                                        f.write(asciiFOTO)
+                                        
+                                    repetir = 'n'
                             
                             sql = 'INSERT INTO candidatos (nome_completo, numero_votacao, nome_partido, foto_ascii) VALUES (%s, %s, %s, %s)'
                             val = (name, number, partido, pastaASCII)
@@ -127,7 +187,7 @@ def gestao_candidatos(conn):
                     def buscar_candidato(conn):
                         try:
                             cursor = conn.cursor()
-                            nome = str(input("\nDigite o nome do candidato: "))
+                            numeroPartido = int(input("\nDigite o numero do partido do candidato: "))
                             sql = 'SELECT * FROM candidatos'
                             cursor.execute(sql)
                             result = cursor.fetchall()
@@ -137,7 +197,7 @@ def gestao_candidatos(conn):
                             print('-' * 120)
                             
                             for candidato in result:
-                                if nome.lower() == candidato[1].lower():
+                                if numeroPartido == candidato[2]:
                                     encontrado = True
                                     
                                     if len(candidato) > 4 and candidato[4]:
